Commented out on_ready message
Docker seems to be restarting container much more frequently then before, the 'yo we in there' message is becoming disruptive. It is also not affecting functionality, so we will comment it out. We leave the on_ready message commented because it is useful for dev versions to know when restarts have happened.